add serialization for  new packets and change for some existing packets switch to using ints instead of floats where relevant

+++ PacketTechnicalData.h
@@ -12,9 +12,9 @@
 #include "Packet.h"
 
 struct PacketTechnicalData {
-	float32_t vcc;
-	float32_t carBatteryVoltage;
-	float32_t carBatteryCurrent;
+	uint16_t vcc; // mV
+	uint16_t carBatteryVoltage; // mV
+	uint16_t carBatteryCurrent; // mA
 	uint16_t freeRAM;
 };
